Extract reusable 'MediaStoreSpecs' helper methods into a shared scope so it can be used by other 'MediaStore' implementation tests
